version: 0.8.0; move NAME_STATES to constant
