+++ app/constants/names.js
@@ -1,0 +1,8 @@
+export const NAME_STATES = {
+  OPENING: 'OPENING',
+  BIDDING: 'BIDDING',
+  REVEAL: 'REVEAL',
+  CLOSED: 'CLOSED',
+  REVOKED: 'REVOKED',
+  TRANSFER: 'TRANSFER',
+};
