+++ package.json
@@ -2,7 +2,7 @@
   "name": "bob-wallet",
   "productName": "Bob",
   "author": "Kyokan Group, Inc.",
-  "version": "0.7.1",
+  "version": "0.8.0",
   "description": "A Handshake wallet and auction manager.",
   "scripts": {
     "build": "./scripts/package.sh",
@@ -115,7 +115,7 @@
     "connected-react-router": "5.0.1",
     "cross-env": "5.2.0",
     "css-loader": "1.0.1",
-    "electron": "9.1.1",
+    "electron": "^9.1.1",
     "electron-builder": "22.10.5",
     "electron-devtools-installer": "2.2.4",
     "enzyme": "^3.8.0",
